% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.clean.R
\name{af.clean}
\alias{af.clean}
\title{Cleans the Afrobarometer dataset.}
\usage{
af.clean(afro)
}
\value{
A clean version of the global dataframe \code{afro}
}
\description{
Normally called by default by \code{af.load()}
}
\examples{
af.clean() # Cleans the global version of Loads the 6th round of the Afrobarometer from local file if available, from the websiteif not
af.load(5, T) # Downloads the 5th round of the Afrobarometer, ignoring local files
}
