% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.code.R
\name{af.code}
\alias{af.code}
\title{Prints out simple Afrobarometer codebook}
\arguments{
\item{afro}{Afrobarometer datafram}

\item{question}{Optional: Specify the question whose description you require}
}
\value{
Returns the descriptions of each variable
}
\description{
Prints out simple Afrobarometer codebook
}
\examples{

# Prints out the codebook for all variables in the current round
# or from the website if the local file does not exist
af.code()

}
