% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.rem.miss.R
\name{af.rem.miss}
\alias{af.rem.miss}
\title{Removes missing, don't know, refused, from afrobarometer data}
\usage{
af.rem.miss(afro, question, silent = FALSE)
}
\arguments{
\item{afro}{An afrobarometer dataframe}

\item{question}{The question that you want to clean}
}
\value{
A table object of the proportion table or cross-tab
}
\description{
Removes missing, don't know, refused, from afrobarometer data
}
\examples{
# Remove all missing responses from a question
afro \%>\% af.tab("q5")

}
