% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.tab.R
\name{af.tab}
\alias{af.tab}
\title{One and two-way tables for categorical questions}
\arguments{
\item{afro}{An afrobarometer dataframe}

\item{question}{A string specifying the dependent varirable to examine}

\item{question2}{A string specifying the independent variable to examine}

\item{miss}{A boolean specifying whether or not to include missing/don't know/refused responses (default FALSE)}

\item{digits}{An integer specifying number of digits to round to (default is 2)}

\item{pos}{For two variables, specifies the style of the barplot: "stack" or "dodge" (default)}
}
\value{
A table object of the proportion table or cross-tab
}
\description{
Easily produced one and two-way tables for categorical questions accompanied by a barplot
}
\details{
If a cross-tab, also performs a chi-square test
}
\examples{
# For question 5: Produce a table of percentages, a table of raw numbers,
# and a barplot of responses, dropping missing/don't know/refused responses
afro \%>\% af.tab("q5")

# For question 5: Produces a table of percentages, a table of raw numbers,
# and a barplot of responses, including missing/don't know/refused responses
# afro \%>\% af.tab("q5", miss=T)

# For questions 5 & 7: Produces a cross-tab table of proportions,
# a table of raw numbers, and a side-by-side barplot of responses,
# dropping missing/don't know/refused responses
afro \%>\% af.tab("q5", "q7")

# For questions 5 & 7: Produces a cross-tab table of proportions,
# a table of raw numbers, and a stacked barplot of responses,
# dropping missing/don't know/refused responses
afro \%>\% af.tab("q5", "q7", pos="stack")
}
