% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.clean.R
\name{af_clean}
\alias{af_clean}
\title{Cleans the Afrobarometer dataset}
\usage{
af_clean(afro)
}
\arguments{
\item{afro}{A dataframe containing Afrobarometer data}
}
\value{
A clean version of the passed in dataframe \code{afro}
}
\description{
Returns a clean version of the Afrobarometer dataset. Normally called by default by \code{af.load()}
}
\examples{

# Returns a cleaned version of the afro dataframe loaded
# using \\code{af.load}
afro \%>\% af.clean()
}
