% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.contains.R
\name{af_contains}
\alias{af_contains}
\title{Returns whether or not a question, or vector of questions, exists in the dataset}
\usage{
af_contains(afro, question)
}
\arguments{
\item{afro}{A dataframe containing Afrobarometer data}

\item{question}{A vector of questions}
}
\value{
A single boolean that is TRUE only if all questions exist in the dataset
}
\description{
Returns whether or not a question, or vector of questions, exists in the dataset
}
\examples{

# Check to see if question 5 exists
afro \%>\% af_contains("q5")

#' # Check to see if questions 5 and 30 exist
afro \%>\% af_contains(c("q5", "q30"))
}
