% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afrobar.R
\docType{package}
\name{afrobar}
\alias{afrobar}
\alias{afrobar-package}
\title{afrobar: A package for simply loading and exploring afrobarometer data}
\description{
The afrobar package provides functions for the simple loading and exploration of afrobarometer data
}
