% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.load.R
\name{af.load}
\alias{af.load}
\title{Loads the Afrobarometer dataset}
\arguments{
\item{round}{A number specifying which round of the Afrobarometer to load (6 by default)}

\item{force}{A boolean specifiying whether or not to force the download of a new copy from the website (false by default)}

\item{clean}{A boolean specifying whether or not to clean the dataset (TRUE by default)}
}
\value{
Returns te dataframe \code{afro}, and sets a global integer \code{af.round}
}
\description{
Loads the Afrobarometer dataset
}
\examples{

# Load the 6th round of the Afrobarometer from local file if available,
# or from the website if the local file does not exist
af.load()

# Download and load the 5th round of the Afrobarometer, ignoring and
# overwriting any local files that exist
af.load(5, force=T)

# Download and load the 3rd round of the Afrobarometer, ignoring and
# overwriting any local files that exist, do not automatically clean
af.load(5, force=T, clean=F)
}
