% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af.load.R
\name{af.load}
\alias{af.load}
\title{Loads Afrobarometer dataset.}
\arguments{
\item{round}{A number specifying which round of the Afrobarometer to load (6 by default).}

\item{force}{A boolean specifiying whether or not to force the download of a new copy from the website (false by default)}
}
\value{
The global dataframe \code{afro},and a global integer \code{af.round}.
}
\description{
Loads Afrobarometer dataset.
}
\examples{
af.load() # Loads the 6th round of the Afrobarometer from local file if available, from the websiteif not
af.load(5, T) # Downloads the 5th round of the Afrobarometer, ignoring local files
}
